% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSTING_JSON.R
\name{COSTING_JSON_05}
\alias{COSTING_JSON_05}
\title{A Cat Function}
\usage{
COSTING_JSON_05(data = ModelData(), level1, level2, level3, level4, level5,
  plotoutput = c("cost", "cost", "cost", "cost", "cost"),
  chart.order = NULL, plot.by = "allocations")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
COSTING_JSON_05()
}
\keyword{cats}
