% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSTING_JSON.R
\name{EmptyDrill}
\alias{EmptyDrill}
\title{A Cat Function}
\usage{
EmptyDrill()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
EmptyDrill()

}
\keyword{cats}
