% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COSTING_JSON.R
\name{sort_LVLs_by}
\alias{sort_LVLs_by}
\title{A Cat Function}
\usage{
sort_LVLs_by(LVL, level, sort.on = "cost", chart.order = NULL)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
sort_LVLs_by()
}
\keyword{cats}
